+++ client-common/src/main/java/org/realityforge/replicant/client/transport/RequestAction.java
@@ -1,0 +1,8 @@
+package org.realityforge.replicant.client.transport;
+
+import javax.annotation.Nullable;
+
+public interface RequestAction
+{
+  void invokeReqest( @Nullable ClientSession session, @Nullable RequestEntry request );
+}